+++ python/test/extras/test_canonical_noise.py
@@ -100,7 +100,7 @@ def test__tulap_cdf_array_input():
     t_values = np.array([-1.0, -0.5, 0.0, 0.5, 1.0])
 
     ptulap_results = _cnd_cdf(t_values, shift=0.0, d_in=1.0, d_out=(0.5, 1e-7))
-    assert list(ptulap_results) == [
+    assert ptulap_results.tolist() == [
         0.30326526920325075,
         0.37754063104407853,
         0.5,
