+++ rust/src/measurements/make_private_expr/expr_noisy_max/test.rs
@@ -37,7 +37,7 @@ fn test_noisy_max_udf() -> Fallible<()> {
     let actual = super::noisy_max_udf(
         &[scores],
         NoisyMaxPlugin {
-            distribution: TopKDistribution::Gumbel,
+            replacement: true,
             negate: false,
             scale: 0.0,
         },
