--- rust/src/measurements/noisy_top_k/exponential/references.bib
@@ -1,24 +1,0 @@
-@article{ding2021permute,
-  title   = {The permute-and-flip mechanism is identical to report-noisy-max with exponential noise},
-  author  = {Ding, Zeyu and Kifer, Daniel and Steinke, Thomas and Wang, Yuxin and Xiao, Yingtai and Zhang, Danfeng and others},
-  journal = {arXiv preprint arXiv:2105.07260},
-  year    = {2021}
-}
-
-@article{mckenna2020permute,
-  title   = {Permute-and-flip: A new mechanism for differentially private selection},
-  author  = {McKenna, Ryan and Sheldon, Daniel R},
-  journal = {Advances in Neural Information Processing Systems},
-  volume  = {33},
-  pages   = {193--203},
-  year    = {2020}
-}
-
-@article{canonne2020discrete,
-  title   = {The discrete gaussian for differential privacy},
-  author  = {Canonne, Cl{\'e}ment L and Kamath, Gautam and Steinke, Thomas},
-  journal = {Advances in Neural Information Processing Systems},
-  volume  = {33},
-  pages   = {15676--15688},
-  year    = {2020}
-}
