--- rust/src/measurements/noisy_top_k/gumbel/references.bib
@@ -1,32 +1,0 @@
-@article{Durfee2019PracticalDP,
-  author       = {David Durfee and
-                  Ryan Rogers},
-  title        = {Practical Differentially Private Top-k Selection with Pay-what-you-get
-                  Composition},
-  journal      = {CoRR},
-  volume       = {abs/1905.04273},
-  year         = {2019},
-  url          = {http://arxiv.org/abs/1905.04273},
-  eprinttype    = {arXiv},
-  eprint       = {1905.04273},
-  timestamp    = {Mon, 02 Mar 2020 11:27:14 +0100},
-  biburl       = {https://dblp.org/rec/journals/corr/abs-1905-04273.bib},
-  bibsource    = {dblp computer science bibliography, https://dblp.org}
-}
-
-@article{Dong2019OptimalDP,
-  author       = {Jinshuo Dong and
-                  David Durfee and
-                  Ryan Rogers},
-  title        = {Optimal Differential Privacy Composition for Exponential Mechanisms
-                  and the Cost of Adaptivity},
-  journal      = {CoRR},
-  volume       = {abs/1909.13830},
-  year         = {2019},
-  url          = {http://arxiv.org/abs/1909.13830},
-  eprinttype    = {arXiv},
-  eprint       = {1909.13830},
-  timestamp    = {Mon, 02 Mar 2020 11:27:19 +0100},
-  biburl       = {https://dblp.org/rec/journals/corr/abs-1909-13830.bib},
-  bibsource    = {dblp computer science bibliography, https://dblp.org}
-}
