+++ rust/src/measurements/noisy_top_k/noisy_top_k.tex
@@ -1,13 +1,13 @@
 \documentclass{article}
-\input{../../../lib.sty}
+\input{../../lib.sty}
 
-\title{\texttt{fn exponential\_top\_k}}
+\title{\texttt{fn noisy\_top\_k}}
 \author{Michael Shoemate}
 
 \begin{document}
 \maketitle\contrib
-This document proves soundness of \rustdoc{measurements/noisy_top_k/exponential/fn}{exponential\_top\_k} in \asOfCommit{mod.rs}{e62b0aa2}. 
-\texttt{exponential\_top\_k} noisily selects the index of the best score from a vector of input scores $k$ times without replacement.
+This document proves soundness of \rustdoc{measurements/noisy_top_k/fn}{noisy\_top\_k} in \asOfCommit{mod.rs}{e62b0aa2}. 
+\texttt{noisy\_top\_k} noisily selects the index of the best score from a vector of input scores $k$ times.
 
 \section{Hoare Triple}
 \subsection*{Preconditions}
@@ -21,16 +21,17 @@ This document proves soundness of \rustdoc{measurements/noisy_top_k/exponential/
 
 \subsection*{Pseudocode}
 \label{sec:python-pseudocode}
-\lstinputlisting[language=Python,firstline=2,escapechar=`]{./pseudocode/exponential_top_k.py}
+\lstinputlisting[language=Python,firstline=2,escapechar=`]{./pseudocode/noisy_top_k.py}
 
 \subsection*{Postcondition}
 \begin{theorem}
     \label{postcondition}
+
     \begin{itemize}
-        \item Returns the index of the top element $z_i$,
-        where each $z_i \sim \mathrm{Exp}(\mathrm{shift}=y_i, \mathrm{scale}=\texttt{scale})$,
-        and each $y_i = -x_i$ if \texttt{negate}, else $y_i = x_i$,
-        $k$ times with removal.
+        \item If replacement is set, returns a sample from $\mathcal{M}_{EM}$ (as defined in MS2023 Definition 4),
+            otherwise returns a sample from $\mathcal{M}_{PF}$ (as defined in MS2023 Lemma 1),
+            $k$ times by peeling,
+            where $\texttt{scale} = \frac{2 \cdot \Delta}{\epsilon}$.
         \item Errors are data-independent, except for exhaustion of entropy.
     \end{itemize}
 \end{theorem}
