+++ rust/src/measurements/noisy_top_k/pseudocode/TopKMeasure_ZeroConcentratedDivergence.py
@@ -1,7 +1,6 @@
 # type: ignore
 # ZeroConcentratedDivergence
-def noisy_top_k(x: list[TIA], scale: f64, k: usize, negate: bool) -> list[usize]:
-    return gumbel_top_k(x, scale, k, negate)
+REPLACEMENT = True
 
 def privacy_map(d_in: f64, scale: f64) -> f64:
     return d_in.inf_div(scale).inf_powi(ibig(2)).inf_div(8.0)
