+++ rust/src/measurements/noisy_top_k/references.bib
@@ -1,19 +1,3 @@
-@article{Durfee2019PracticalDP,
-  author     = {David Durfee and
-                Ryan Rogers},
-  title      = {Practical Differentially Private Top-k Selection with Pay-what-you-get
-                Composition},
-  journal    = {CoRR},
-  volume     = {abs/1905.04273},
-  year       = {2019},
-  url        = {http://arxiv.org/abs/1905.04273},
-  eprinttype = {arXiv},
-  eprint     = {1905.04273},
-  timestamp  = {Mon, 02 Mar 2020 11:27:14 +0100},
-  biburl     = {https://dblp.org/rec/journals/corr/abs-1905-04273.bib},
-  bibsource  = {dblp computer science bibliography, https://dblp.org}
-}
-
 @article{Dong2019OptimalDP,
   author     = {Jinshuo Dong and
                 David Durfee and
@@ -41,9 +25,12 @@
   url           = {https://arxiv.org/abs/2010.12603}
 }
 
-@article{ding2021permute,
-  title   = {The permute-and-flip mechanism is identical to report-noisy-max with exponential noise},
-  author  = {Ding, Zeyu and Kifer, Daniel and Steinke, Thomas and Wang, Yuxin and Xiao, Yingtai and Zhang, Danfeng and others},
-  journal = {arXiv preprint arXiv:2105.07260},
-  year    = {2021}
+@misc{cesar2020bounding,
+  title         = {Bounding, Concentrating, and Truncating: Unifying Privacy Loss Composition for Data Analytics},
+  author        = {Mark Cesar and Ryan Rogers},
+  year          = {2020},
+  eprint        = {2004.07223},
+  archiveprefix = {arXiv},
+  primaryclass  = {cs.CR},
+  url           = {https://arxiv.org/abs/2004.07223}
 }
