+++ rust/src/traits/samplers/cks20/mod.rs
@@ -78,7 +78,6 @@ fn sample_bernoulli_exp1(x: RBig) -> Fallible<bool> {
 /// For any non-negative finite rational `x`,
 /// `sample_bernoulli_exp` either returns `Err(e)` due to a lack of system entropy,
 /// or `Ok(out)`, where `out` is distributed as $Bernoulli(exp(-x))$.
-
 pub(crate) fn sample_bernoulli_exp(mut x: RBig) -> Fallible<bool> {
     // Sample floor(x) independent Bernoulli(exp(-1))
     // If all are 1, return Bernoulli(exp(-(x-floor(x))))
